+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
+## 0.0.1-dev.1
 
-* TODO: Describe initial release.
+* Initial pre-release version.
+* Added basic HTTP operations in `HttpOperationsRepository`.
+* Implemented `HttpException` for better error handling.
+* [Add any other notable changes or features here]